pyproject.toml
[tool.poetry]
name = "commonroad-dataset-converter"
version = "2025.1.0"
description = "CommonRoad Dataset Converter"
authors = ["Cyber-Physical Systems Group, Technical University of Munich <commonroad@lists.lrz.de>"]
license = "BSD-3-Clause"
readme = "README.md"
packages = [{ include = "crdata" }]
homepage = "https://commonroad.in.tum.de"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
exclude = ["**/*.ipynb"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = "^1.21"
scipy = ">=1.4.1"
pandas = ">=0.24.2"
ruamel-yaml = ">=0.16.10"
commonroad-io = "^2024.2"
typer = ">=0.9,<0.13"
click = "<=8.1"
tqdm = "^4.64.1"
pyproj = "^3.4.0"
pyarrow = ">=10.0.1"
commonroad-route-planner = "^2025.1.0"
commonroad-drivability-checker = "^2025.3.1"
commonroad-clcs = "^2025.2.0"
pydantic = "^2.3.0"

[tool.poetry.scripts]
crconvert = 'crdata.main:cli'

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
pytest = "^8.3.2"
commonroad-scenario-designer = { git = "https://github.com/CommonRoad/commonroad-scenario-designer", branch = "develop" }

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.0"
mkdocstrings-python = "^1.8.0"
mkdocs-material = "^9.5.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
